// Client/projects/fibi/src/app/admin-modules/person-maintenance/person-manitenance.interface.ts
export interface Person {
    addressLine1?: string;
    addressLine2?: string;
    addressLine3?: string;
    age?: number | string;
    city?: string;
    country?: any;
    countryDetails?: any;
    countryOfCitizenshipDetails?: any;
    countryOfCitizenshipCode?: any;
    countryCode?: string;
    dateOfBirth?: number | Date | string;
    dateOfInactive?: number | Date | string;
    directoryDepartment?: string | number;
    directoryTitle?: string;
    educationLevel?: string;
    emailAddress?: string;
    faxNumber?: number | string;
    firstName?: string;
    fullName?: string;
    gender?: string;
    homeUnit?: string | number;
    isExternalUser?: boolean;
    isFaculty?: boolean;
    isGraduateStudentStaff?: boolean;
    isMedicalStaff?: boolean;
    isOtherAcadamic?: boolean;
    isPasswordChange?: boolean;
    isResearchStaff?: boolean;
    isServiceStaff?: boolean;
    isSupportStaff?: boolean;
    isUsernameChange?: boolean;
    isWebhookActive?: boolean;
    lastName?: string;
    middleName?: string;
    mobileNumber?: string;
    officeLocation?: string;
    officePhone?: string;
    orcidId?: number | string;
    pagerNumber?: number | string;
    personId?: string | number;
    postalCode?: number | string;
    primaryTitle?: string;
    principalName?: string;
    priorName?: string;
    salaryAnniversary?: number | Date | string;
    school?: string;
    secOfficeLocation?: string;
    secOfficePhone?: number | string;
    state?: string;
    status?: string;
    supervisorPersonId?: string | number;
    unit?: Unit;
    updateTimestamp?: number;
    updateUser?: string;
    password?: string;
    jobCode?: string;
}

export interface Unit {
    unitNumber?: string;
    parentUnitNumber?: string;
    organizationId?: any;
    unitName?: string;
    active?: boolean;
    updateTimestamp?: number;
    updateUser?: string;
    acronym?: any;
    isFundingUnit?: any;
    unitAdministrators?: any[];
    unitDetail?: string;
    parentUnitName?: any;
    organizationName?: any;
}